// src/app/@api/model/invoiceDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DealStepDto } from './dealStepDto';
import { InquiryVersionIdDto } from './inquiryVersionIdDto';
import { OrderIdDto } from './orderIdDto';
import { QuotationVersionIdDto } from './quotationVersionIdDto';


export interface InvoiceDto { 
    comment?: string;
    dealStep?: DealStepDto;
    inquiryVersion?: InquiryVersionIdDto;
    invoiceId?: number;
    order?: OrderIdDto;
    quotationVersion?: QuotationVersionIdDto;
    status?: number;
}